backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
spacy==3.7.2
python-dateutil==2.8.2
python-multipart==0.0.6
requests==2.31.0

# For SpaCy English model
# Run: python -m spacy download en_core_web_md